check if the bot can send/connect to (voice)channels
Besides various minor adjustments, the bot will now check if it can send
to the channel a command is called in. If it cannot, the message author
will be directly messaged. If it can, proceed as normal.

If it cannot connect to a voice channel, it will directly message the
person that used the command of this.

Using a command that uses voice channels from within a text channel the
bot can read, but not send in, will make the command progress as
expected, as long as no messages need to be sent.