Make some commands usable by admins/mods
Now, after seperating the ignore lists to a per-server list, everyone on
a discord server with the kick or ban permission will be able to use the
ignore command.

Also, users with the kick or ban permission will be able to set the
bot's nickname for their server.

setName command has been renamed to setUsername, new command is
setNickname.

All commands have been adjusted to go along with the change in
permission-checking (permission -> botPerm, new var is userPerm).

The bot owner retains the ability to use the ignore and setNickname
command on any server, no matter if he  has the kick/ban permission or
not.