Add command to reload loaded commands
You can now reload a chosen command by passing the bot the "reload"
command and your chosen command as argument.

Example usage would be "robbot, reload about" to reload the about
command.

In addition, the help command is always reloaded, so that if new
commands are added, removed or changed, the output of the help command
will stay up-to-date.

It is not possible to reload all commands. To do this, restart the bot.