Make reload argument case-insensitive
You can now pass the reload command case-insensitive commands and it
will reload it successfully.

Example: "robbot, reload HeLP" will reload the 'help' command, and not
report an error about not being able to find the module anymore.

Invalid (not found, typo'd) commands passed as argument will report an
error to the user.

Also make the showLog command tell the user which logs are available,
instead of just reporting that the passed log is invalid.