Finalize toggle command plus adjustments
The toggle command has been finalized, in accordance with the previous
commit.

The help command will now include a list of commands disabled on the
server the command is called on.

Toggle, help and any commands that are invalid are NOT toggleable.

There have also been adjustments made to some comments and some code
parts.